Split SSL verification into new source file
The SSL verification is used by TLS and DTLS, refactor to a separate
translation unit.
