+++ src/netlog/netlog-ssl.h
@@ -1,0 +1,6 @@
+/* SPDX-License-Identifier: LGPL-2.1-or-later */
+#pragma once
+
+#include <openssl/ssl.h>
+
+int ssl_verify_certificate_validity(int status, X509_STORE_CTX *store);
